*moved all timecode code in the timecode plugin *renamed backend to protocol *removed hRes and enabled from app settings
