+++ lisp/default.cfg
@@ -70,8 +70,5 @@ dbMin = -60
 dbClip = 0
 
 [Timecode]
-Enabled = True
-Format = FILM
-HRes = True
-Backend = MIDI
-
+Format = SMPTE
+Protocol = MIDI
