--- lisp/modules/timecode/__init__.py
@@ -1,1 +1,0 @@
-from  .timecode import Timecode
